Add the kernel build system
The build system consists of a very simple three-tier Makefile
scheme. At the root of the source tree there is the top level
Makefile. The kernel Makefile (located in scripts/Makefile.kernel)
defines build routines specific to the kernel, and calls the third level
of Makefile (located in arch/<arch>/Makefile.<arch>) which defines the
build configuration specific to that particular architecture.