Define a DEBUG macro
This generic build-time define can be used to guard the compilation of
tests and assertions which are not required for release builds, but are
useful for testing and debugging purposes.

As the project develops, it may be beneficial to introduce a more fine
grained debug build approach than simply an on/off switch, this could be
implemented by breaking up the debugging flag into smaller
subflags (BOOT_DEBUG, STACK_DEBUG, etc) and enabling those through a
debug level control (DEBUG=0-4, for example).