build: Link asm objects before C
GRUB looks for the Multiboot header in the first 8kb of the kernel. If
it does not find it, it throws an error 13. The multiboot header is
normally defined in an assembly file (e.g. arch/x86/boot/boot.s). To
increase the chances that the multiboot header is included in the
correct place, link the asm files before the C files.

This is essentially a hacky solution to the problem. A better solution
would be to define the multiboot header in the linker script itself.

See: http://wiki.osdev.org/Grub_Error_13