tools: Add a headercheck script
This script generates a list of all header files, and then iterates over
each header and source file, listing those which include the header. The
idea here is that it provides a convenient means for checking to see how
widely used a source file is. For example, this tool could be used as a
reference before changing a function prototype, to see how many source
files will be affected.